refactor(drift_sync): extract issue_state.py (7/8)
Move parse_zero_findings_timestamps, should_close_issue, and
resolve_drift_issue into issue_state.py. Top of the DAG — depends on
formatters for issue body/comment rendering and on github_api.issues
for create/update/close.

Unused imports (re, datetime, timedelta, Any) cleaned from __init__.py.
At this point __init__.py contains only re-exports + the doctor bridge
side-effect import.

Refs #47 (Theme A item 4).
